+++ finagle-core/src/main/scala/com/twitter/finagle/loadbalancer/TrafficDistributor.scala
@@ -78,6 +78,7 @@ private[finagle] object TrafficDistributor {
   private class Distributor[Req, Rep](
     classes: Iterable[WeightClass[Req, Rep]],
     busyWeightClasses: Counter,
+    reuseEndpoints: Boolean,
     rng: Rng = Rng.threadLocal)
       extends ServiceFactory[Req, Rep] {
 
@@ -136,13 +137,18 @@ private[finagle] object TrafficDistributor {
         }
       }
 
-    def close(deadline: Time): Future[Unit] =
-      Closable
-        .all(
-          Closable.all(balancers: _*),
-          Closable.all(endpoints: _*)
-        )
-        .close(deadline)
+    def close(deadline: Time): Future[Unit] = {
+      if (reuseEndpoints) {
+        Closable.all(balancers: _*).close(deadline)
+      } else {
+        Closable
+          .all(
+            Closable.all(balancers: _*),
+            Closable.all(endpoints: _*)
+          )
+          .close(deadline)
+      }
+    }
 
     private[this] val svcFactoryStatus: ServiceFactory[Req, Rep] => Status =
       sf => sf.status
@@ -169,6 +175,7 @@ private[finagle] object TrafficDistributor {
 private class TrafficDistributor[Req, Rep](
   dest: Event[Activity.State[Set[AddressedFactory[Req, Rep]]]],
   newBalancer: (Activity[Set[EndpointFactory[Req, Rep]]], Boolean) => ServiceFactory[Req, Rep],
+  reuseEndpoints: Boolean,
   rng: Rng = Rng.threadLocal,
   statsReceiver: StatsReceiver = NullStatsReceiver)
     extends ServiceFactory[Req, Rep] {
@@ -285,7 +292,7 @@ private class TrafficDistributor[Req, Rep](
   private[this] val underlying: Event[ServiceFactory[Req, Rep]] =
     weightClasses.foldLeft(init) {
       case (_, Activity.Ok(wcs)) =>
-        val dist = new Distributor(wcs, busyWeightClasses, rng)
+        val dist = new Distributor(wcs, busyWeightClasses, reuseEndpoints, rng)
         updateGauges(wcs)
         pending.updateIfEmpty(Return(dist))
         dist