+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+Bug Fixes
+~~~~~~~~~~
+
+* finagle-core/partitioning: Close balancers and their gauges when repartitioning.
+  ``PHAB_ID=D731675``
+
 Runtime Behavior Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~